v1: move all oscap related handlers and helpers to handler_oscap
